requirements.txt
asgiref==3.9.1
Django==5.2.5
django-admin-interface==0.31.0
django-colorfield==0.11.0
gunicorn==23.0.0
google-genai==1.51.0
packaging==25.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing-extensions==4.14.1
whitenoise==6.9.0
pymysql==1.1.1
scikit-learn>=1.3.0
pandas>=2.0.0
joblib>=1.3.0
